backend/requirements.txt
fastapi
uvicorn
pandas
pdfplumber
openpyxl
python-multipart
google-generativeai